requirements.txt
GitPython==3.1.1
Jinja2==2.11.2
